move model equations to separate module
